Updating Jinja filters to conform to Ansible 2.5+
Since Ansible 2.5, the use of jinja tests as filters has been deprecated.
I've run the script provided by the ansible team to 'fix' the jinja filters
to conform to the newer syntax.

This fixes the deprecation warnings.

This is a specific run of the script against active stein files, rather
than a cherry pick of 656767.

Change-Id: I33f486c4cc2990d270821d70c73e9654004e90e0