(Rocky and Queens only) CI: Fix ceph jobs for kolla
The way Zuul handles job dependencies caused the computed path
to setup_ceph_disks.sh to be wrong when job was run due to kolla
(as opposed to kolla-ansible) change.

This patch uses a relative path as is the case in Stein and later.

Change-Id: I4f45cd18b86ebe0efcc3101e31e4bcafdc83eb57
Signed-off-by: Radosław Piliszek <radoslaw.piliszek@gmail.com>
(cherry picked from commit 1d22714cfeb5bdde43255286549ea781d67f52dd)