Test minimum supported and latest versions of Ansible
We have a minimum supported version of Ansible, currently 2.4. We should
test this in addition to the latest version. This change tests latest on
Ubuntu, and minimum on other distros.

Also limited ARA to 0.15, as necessary to work with Ansible 2.4.

Change-Id: I45a7173139f057177a71e919ad3e718a99d9f87b
(cherry picked from commit b9da05be556ad8fb847d968a7240557829c5bfdb)