Use default BGP speaker driver in bgp_dragent.ini
In the Stein release, neutron-dynamic-routing switched from
using ryu to os-ken (a fork of ryu). The config option
bgp_speaker_driver [1] is currently set in kolla-ansible to use the ryu
driver, but this is not included in the image due to it being removed
from neutron-dynamic-routing's requirements.txt.

This change uses the os-ken driver for bgp_speaker_driver, since this
package is included in the image.

[1] https://docs.openstack.org/neutron-dynamic-routing/latest/contributor/dragent-drivers.html#bgp-driver

Change-Id: I76ad84c6b963818208f080ebeb054b2553ff40af
Closes-Bug: #1839165