Don't stop mariadb slaves if they aren't running
Stein only

If mariadb is not running on all of the slave nodes during a deploy,
reconfigure or upgrade, then currently the "remove restart policy from
slave mariadb" handler will fail on nodes where the container does not
exist. A valid use case for this is when scaling out the cluster, but it
could also happen if the slaves are down for some other reason.

This change checks if the mariadb container exists on slaves, and skips
all shutdown related tasks on slaves where it does not.

Change-Id: I9a191da0ec6ac13ec6dd882b1723d3686f9edc8c
Closes-Bug: #1840764