Don't assume etcd group exists in baremetal role
The baremetal role does not currently assume too much about the
inventory, and in kayobe the seed is deployed using a very minimal
inventory.

Icf3f01516185afb7b9f642407b06a0204c36ecbe added a reference to the etcd
group in the baremetal role, which causes kayobe seed deployment to fail
with the following error:

    AnsibleUndefinedVariable: 'dict object' has no attribute 'etcd'

This change defaults the group lookup to an empty list.

Change-Id: Ib3252143a97652c5cf70b56cbfd7c7ce69f93a55
Closes-Bug: #1841073