Fix kolla_toolbox error handling
Backport only.

The kolla_toolbox Ansible module executes as-hoc ansible commands in the
kolla_toolbox container, and parses the output to make it look as if
ansible-playbook executed the command. Currently however, this module
sometimes fails to catch failures of the underlying command, and also
sometimes shows tasks as 'ok' when the underlying command was changed.

This change introduces a check on the status field for non-JSON output,
and assumes the module has changes on success.

NOTE: The original patch on master
(https://review.opendev.org/#/c/682340/) cannot be backported as it
relies on Ansible 2.5+ in the kolla_toolbox container.

Change-Id: Ib1e33f8e5c60d98b76a4eba9f405af89431d60bc
Closes-Bug: 1844114