[neutron] Adjust neighbour table thresholds
When clouds have a large number of hosts, the default size of the ARP
cache is too small. The cache can overflow, which means that the system
has no way to reach some IP addresses.

Increasing threshold limits addresses the situation, in a reasonably
safe way (the maximum impact is 5MB or so of additional RAM used).

More context on this issue:

* http://man7.org/linux/man-pages/man7/arp.7.html
* https://bugs.launchpad.net/charm-nova-compute/+bug/1780348
* https://bugs.launchpad.net/fuel/+bug/1488938
* https://bugs.launchpad.net/tripleo/+bug/1690087
* https://github.com/crowbar/crowbar-openstack/commit/0583a0c94996df6b784229e8a534f955eaca85bc
* https://github.com/crowbar/crowbar-openstack/commit/3dd21ea62ac152e40bfdfee4b8e25a528c82a79f
* https://opendev.org/openstack/tripleo-heat-templates/commit/1651a1805a16212299fe0a91aebb2a91ed39bc6e

Change-Id: I60c871e8eb9f2c086818ff077987f2390930800c
Closes-Bug: #1844349