Fix fernet_token_expiry precheck
In the case of 3600, the issue was that the regex `360` was matching
`3600` leaving `0` which failed the regex search. We could reverse the
order of the regexs so that longer ones appeared before shorter ones,
but why not just check membership in list?

Change-Id: I089f1e288bae6c093be66181c81a4373a6ef3de4
Closes-Bug: #1856021
(cherry picked from commit 951790e103c8d77ca13da6df685df38501e8f512)