Fix nova and cinder handlers with no ceph
Stein only

If we deploy nova or cinder without internal or external ceph
configured, and there are no config or container changes, then the
following task may fail:

RUNNING HANDLER [nova : Restart nova-libvirt container]

The error is in evaluating the conditionals:

'str object' has no attribute 'changed'

Thank you to Emanuel Evans (shosti) for determining that ceph_conf is the issue
here. The problem is that if ceph_confs is undefined, then we set ceph_conf to
{{ none }}, however due to the 'magic' of Jinja this evaluates to a string. The
check for none therefore fails, and we look up an attribute 'changed' on a
string.

This change fixes the issue by changing the condition from 'is not none' to 'is
mapping'. We also change the value to an empty string, to make it clear that it
does not end up as None.

Change-Id: I240ecd6b9a4ef09ef9cbc3f58c9c0169c73c5e58
Closes-Bug: #1842913