Fix restart handlers for polling.yaml changes
Changes to polling.yaml were not triggering restart handlers due to a
missing check on polling_overwriting.changed. This is committed straight
to stable/stein, as stable/train and later have removed this code.

The polling_overwriting variable is also removed from the
ceilometer-notification handler as it doesn't use this file.

Change-Id: I1e665c7b276d9c885881d96315daac4fa78682ce
Closes-Bug: #1860651